chat_backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.1
